requirements.txt
asgiref==3.8.1
cachetools==5.5.0
certifi==2024.7.4
chardet==5.2.0
charset-normalizer==3.3.2
Django==5.0.6
docx==0.2.4
et-xmlfile==1.1.0
google-api-core==2.19.2
google-api-python-client==2.142.0
google-auth==2.34.0
google-auth-httplib2==0.2.0
google-auth-oauthlib==1.2.1
googleapis-common-protos==1.65.0
httplib2==0.22.0
idna==3.8
lxml==5.3.0
mysqlclient==2.2.4
oauthlib==3.2.2
openpyxl==3.1.5
packaging==24.1
pillow==10.3.0
proto-plus==1.24.0
protobuf==5.27.4
pyasn1==0.6.0
pyasn1_modules==0.4.0
PyMuPDF==1.24.9
PyMuPDFb==1.24.9
pyparsing==3.1.4
PyPDF2==3.0.1
pytesseract==0.3.13
python-docx==1.1.2
reportlab==4.2.0
requests==2.32.3
requests-oauthlib==2.0.0
rsa==4.9
sqlparse==0.5.0
typing_extensions==4.12.2
tzdata==2024.1
uritemplate==4.1.1
urllib3==2.2.2
